don't show login while waiting for BE
